Added safe patching and dumping with retries
This should be more robust and prevent bricks. Each 16 byte of dump will
retry up to 10 times if it detects the data does not appear to be from
the SPI. There are weird cases that the command falls through and the
input is returned (see #2).

Flashing is now followed by a read-back. If a read-back (with the retry
behaviour above) fails, then it will retry the flash (up to 100 times).